Fix verify user function
